+++ mullvad-daemon/src/lib.rs
@@ -220,6 +220,8 @@ pub enum DaemonCommand {
     SetBridgeState(ResponseTx<(), settings::Error>, BridgeState),
     /// Set if IPv6 should be enabled in the tunnel
     SetEnableIpv6(ResponseTx<(), settings::Error>, bool),
+    /// Set whether to enable PQ PSK exchange in the tunnel
+    SetQuantumResistantTunnel(ResponseTx<(), settings::Error>, bool),
     /// Set DNS options or servers to use
     SetDnsOptions(ResponseTx<(), settings::Error>, DnsOptions),
     /// Toggle macOS network check leak
@@ -979,6 +981,9 @@ where
             }
             SetBridgeState(tx, bridge_state) => self.on_set_bridge_state(tx, bridge_state).await,
             SetEnableIpv6(tx, enable_ipv6) => self.on_set_enable_ipv6(tx, enable_ipv6).await,
+            SetQuantumResistantTunnel(tx, enable_pq) => {
+                self.on_set_quantum_resistant_tunnel(tx, enable_pq).await
+            }
             SetDnsOptions(tx, dns_servers) => self.on_set_dns_options(tx, dns_servers).await,
             SetWireguardMtu(tx, mtu) => self.on_set_wireguard_mtu(tx, mtu).await,
             SetWireguardRotationInterval(tx, interval) => {
@@ -1053,7 +1058,7 @@ where
                 }
             }
             PrivateDeviceEvent::RotatedKey(_) => {
-                if let Some(TunnelType::Wireguard) = self.get_target_tunnel_type() {
+                if self.get_target_tunnel_type() == Some(TunnelType::Wireguard) {
                     self.schedule_reconnect(WG_RECONNECT_DELAY);
                 }
             }
@@ -1683,7 +1688,7 @@ where
                         .set_tunnel_options(&self.settings.tunnel_options);
                     self.event_listener
                         .notify_settings(self.settings.to_settings());
-                    if let Some(TunnelType::Wireguard) = self.get_connected_tunnel_type() {
+                    if let Some(TunnelType::Wireguard) = self.get_target_tunnel_type() {
                         log::info!("Initiating tunnel restart");
                         self.reconnect_tunnel();
                     }
@@ -1834,7 +1839,7 @@ where
                         .set_tunnel_options(&self.settings.tunnel_options);
                     self.event_listener
                         .notify_settings(self.settings.to_settings());
-                    if let Some(TunnelType::OpenVpn) = self.get_connected_tunnel_type() {
+                    if self.get_target_tunnel_type() == Some(TunnelType::OpenVpn) {
                         log::info!(
                             "Initiating tunnel restart because the OpenVPN mssfix setting changed"
                         );
@@ -1950,6 +1955,36 @@ where
             Err(e) => {
                 log::error!("{}", e.display_chain_with_msg("Unable to save settings"));
                 Self::oneshot_send(tx, Err(e), "set_enable_ipv6 response");
+            }
+        }
+    }
+
+    async fn on_set_quantum_resistant_tunnel(
+        &mut self,
+        tx: ResponseTx<(), settings::Error>,
+        use_pq_safe_psk: bool,
+    ) {
+        let save_result = self
+            .settings
+            .set_quantum_resistant_tunnel(use_pq_safe_psk)
+            .await;
+        match save_result {
+            Ok(settings_changed) => {
+                Self::oneshot_send(tx, Ok(()), "set_quantum_resistant_tunnel response");
+                if settings_changed {
+                    self.parameters_generator
+                        .set_tunnel_options(&self.settings.tunnel_options);
+                    self.event_listener
+                        .notify_settings(self.settings.to_settings());
+                    if self.get_target_tunnel_type() == Some(TunnelType::Wireguard) {
+                        log::info!("Reconnecting because the PQ safety setting changed");
+                        self.reconnect_tunnel();
+                    }
+                }
+            }
+            Err(e) => {
+                log::error!("{}", e.display_chain_with_msg("Unable to save settings"));
+                Self::oneshot_send(tx, Err(e), "set_quantum_resistant_tunnel response");
             }
         }
     }