+++ windows/winfw/src/winfw/rules/shared.h
@@ -1,6 +1,9 @@
 #pragma once
 
 #include <vector>
+#include <memory>
+#include <winfw/winfw.h>
+#include <libwfp/conditions/conditionprotocol.h>
 #include <libwfp/ipaddress.h>
 
 namespace rules
@@ -9,5 +12,9 @@ namespace rules
 using IpSet = std::vector<wfp::IpAddress>;
 
 void SplitAddresses(const IpSet &in, IpSet &outIpv4, IpSet &outIpv6);
+
+std::unique_ptr<wfp::conditions::ConditionProtocol> CreateProtocolCondition(WinFwProtocol protocol);
+
+bool ProtocolHasPort(WinFwProtocol protocol);
 
 }